Moved dtos folder to database, removed unnecessary comment.
